finished up most of the backend leaks

+++ gridftp/client/test/globus-ftp-client-run-tests.pl
@@ -116,7 +116,6 @@ sub setup_server()
     my $server_host = "localhost";
     my $server_port = 0;
     my $server_nosec = "";
-    my $server_env = "";
     my $subject;
     if(defined($nogsi))
     {
@@ -131,9 +130,6 @@ sub setup_server()
         {
             $ENV{X509_CERT_DIR} = cwd();
             $ENV{X509_USER_PROXY} = cwd() . "/testcred.pem";
-            $server_env = "env "
-                . "X509_USER_CERT=$ENV{X509_USER_PROXY} "
-                . "X509_USER_KEY=$ENV{X509_USER_PROXY} "
         }
     
         system('chmod go-rw testcred.pem');
@@ -141,8 +137,11 @@ sub setup_server()
         $subject = `grid-proxy-info -identity`;
         chomp($subject);
         
-        $ENV{GRIDMAP}= cwd() . "/gridmap";
-        system('mv $GRIDMAP $GRIDMAP.old');       
+        $ENV{GRIDMAP} = cwd() . "/gridmap";
+        if ( -f $ENV{GRIDMAP})
+        {
+            system('mv $GRIDMAP $GRIDMAP.old');    
+        }   
         if( 0 != system("grid-mapfile-add-entry -dn \"$subject\" -ln `whoami` -f $ENV{GRIDMAP} >/dev/null 2>&1") / 256)
         {
             print "Unable to create gridmap file\n";
@@ -150,7 +149,7 @@ sub setup_server()
         }
     }
 
-    $server_pid = open(SERVER, "$server_env $server_prog $server_args |");
+    $server_pid = open(SERVER, "$server_prog $server_args |");
      
     if($server_pid == -1)
     {
